requirements.txt
django==4.2.7
pandas==2.1.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
tqdm==4.66.1
djangorestframework==3.14.0
ipykernel==6.26.0